Move get_eth2_network_config
