+++ common/clap_utils/Cargo.toml
@@ -12,3 +12,4 @@ hex = "0.4.2"
 dirs = "3.0.1"
 eth2_network_config = { path = "../eth2_network_config" }
 eth2_ssz = "0.4.0"
+ethereum-types = "0.12.1"
